Roughly splits zipfile parser out into separate module
